Move the interceptor interface to correct place
