+++ src/main/java/com/laserfiche/repository/api/RepositoryApiClientInterceptor.java
@@ -1,4 +1,4 @@
-package com.laserfiche.repository.api.clients;
+package com.laserfiche.repository.api;
 
 import kong.unirest.Interceptor;
