+++ src/main/java/com/laserfiche/repository/api/SelfHostedInterceptor.java
@@ -4,7 +4,6 @@ import com.laserfiche.api.client.apiserver.TokenClient;
 import com.laserfiche.api.client.httphandlers.Request;
 import com.laserfiche.api.client.httphandlers.RequestImpl;
 import com.laserfiche.api.client.httphandlers.UsernamePasswordHandler;
-import com.laserfiche.repository.api.clients.RepositoryApiClientInterceptor;
 import kong.unirest.Config;
 import kong.unirest.HttpRequest;
